% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{add_nested_pars}
\alias{add_nested_pars}
\title{Add nested parameters}
\usage{
add_nested_pars(parameter_set, nest)
}
\arguments{
\item{parameter_set}{named list of parameters to be added to}

\item{nest}{named list of parameters}
}
\value{
list
}
\description{
Add nested parameters
}
\details{
Utility function for parsing griddle file contents.
Look at the keys shared between \code{pars} and \code{nested_pars}. If the values
associated with those keys are the same in both lists, then add the
other key-value pairs from \code{nested_pars} into \code{pars}.
}
