% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{could_be_integer}
\alias{could_be_integer}
\title{Either is an integer, or equal to its integer cast}
\usage{
could_be_integer(x)
}
\arguments{
\item{x}{value}
}
\value{
boolean
}
\description{
Either is an integer, or equal to its integer cast
}
