% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_map}
\alias{flatten_map}
\title{Convert a named list of lists into a tibble}
\usage{
flatten_map(x, nm_column = "parameter_hash")
}
\arguments{
\item{x}{named list, where each element is a list that can be coerced to
tibble}

\item{nm_column}{column in the output tibble to assign the list names to}
}
\value{
tibble
}
\description{
Convert a named list of lists into a tibble
}
