% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_named_parameter_sets}
\alias{flatten_named_parameter_sets}
\title{Convert a named list of lists into a tibble}
\usage{
flatten_named_parameter_sets(x, nm_column = "parameter_hash")
}
\arguments{
\item{x}{named list, where each list element is a named list of parameters,
which will be flattened into a tibble}

\item{nm_column}{column in the output tibble to assign the list names to}
}
\value{
tibble
}
\description{
Convert a named list of lists into a tibble
}
