% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_parameter_set}
\alias{flatten_parameter_set}
\title{Convert a parameter set into a tibble}
\usage{
flatten_parameter_set(x)
}
\arguments{
\item{x}{named list}
}
\value{
tibble
}
\description{
Convert a parameter set into a tibble
}
