% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_run}
\alias{flatten_run}
\title{Flatten results and parameters into a single tibble}
\usage{
flatten_run(output, remove_fixed = TRUE)
}
\arguments{
\item{output}{output from \code{\link[=run]{run()}}}

\item{remove_fixed}{boolean; drop parameters that don't vary}
}
\value{
tibble, with results and parameters
}
\description{
Flatten results and parameters into a single tibble
}
