% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_value}
\alias{flatten_value}
\title{Convert a value, potentially a vector, into a single value}
\usage{
flatten_value(x)
}
\arguments{
\item{x}{input value}
}
\value{
length-1 value, whose type depends on \code{x}
}
\description{
Use \code{str()} to convert objects with length greater than 1
into a single string.
}
