% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{parse_griddle}
\alias{parse_griddle}
\title{Convert griddle file contents into a list of parameter sets}
\usage{
parse_griddle(griddle)
}
\arguments{
\item{griddle}{griddle file contents}
}
\value{
list of parameter sets, each of which is a named list
}
\description{
Convert griddle file contents into a list of parameter sets
}
