% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cache.R
\name{query_cache}
\alias{query_cache}
\title{Read from cache}
\usage{
query_cache(path)
}
\arguments{
\item{path}{path to cache}
}
\description{
Read from cache
}
\seealso{
\code{\link[=run_cache]{run_cache()}}
}
