% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{read_griddle}
\alias{read_griddle}
\title{Read a griddle file}
\usage{
read_griddle(path)
}
\arguments{
\item{path}{YAML path}
}
\value{
list, see \code{\link[=parse_griddle]{parse_griddle()}}
}
\description{
Read a griddle file
}
