% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameter_sets.R
\name{read_parameter_sets}
\alias{read_parameter_sets}
\title{Read and validate a parameter sets file}
\usage{
read_parameter_sets(path)
}
\arguments{
\item{path}{path to YAML parameter sets file}
}
\value{
parsed YAML
}
\description{
Read and validate a parameter sets file
}
