% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated.R
\name{replicated}
\alias{replicated}
\title{Functional to run a function many times, with a seed}
\usage{
replicated(fun)
}
\arguments{
\item{fun}{function of a parameter set, that returns a tibble}
}
\value{
"replicated" function
}
\description{
Functional to run a function many times, with a seed
}
\details{
The input \code{fun} must take a parameter set as its first argument. (It may
also take optional arguments.) It must return a tibble (or anything else
that \code{\link[purrr:map_dfr]{purrr::map_dfr()}} can coerce). That tibble must not contain the name
\code{replicate}.

A "replicated" function is returned. The
replicated function has the same function signature. It assumes that the
parameter set has elements \code{seed} and \code{n_replicates}. It removes those
elements from the list, sets the seed, calls \code{fun} on the remaining
parameter set \code{n_replicates} times. Each output is augmented with a column
\code{replicate}, and the outputs are row-bound and returned.
}
\examples{
f <- function(pars) tibble::tibble(x_plus_1 = pars$x + 1)
rep_f <- replicated(f)
rep_f(list(x = 1, seed = 42, n_replicates = 3))
}
