% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a function over each of the parameter sets}
\usage{
run(fun, parameter_sets)
}
\arguments{
\item{fun}{function of a parameter set}

\item{parameter_sets}{list of lists, each of which is a named list of
parameters}
}
\description{
Run a function over each of the parameter sets
}
