% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cache.R
\name{run_cache}
\alias{run_cache}
\title{Run simulation function over many parameter sets}
\usage{
run_cache(fun, parameter_sets, path)
}
\arguments{
\item{fun}{simulation function, takes in a parameter set}

\item{parameter_sets}{list of parameter sets}

\item{path}{path to cache location}
}
\description{
Run simulation function over many parameter sets
}
\details{
Runs, optionally in parallel, using \code{\link[furrr:future_map]{furrr::future_walk()}}.
}
\seealso{
\code{\link[=query_cache]{query_cache()}}
}
