% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{update_list}
\alias{update_list}
\title{Update a list with values from another list}
\usage{
update_list(x, y)
}
\arguments{
\item{x}{starting list}

\item{y}{updating list}
}
\value{
list, with all names present in \code{x} and \code{y}, with values from \code{y}
if a name is present in both
}
\description{
Update a list with values from another list
}
