% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_griddle.R
\name{validate_griddle}
\alias{validate_griddle}
\title{Validate griddle file contents}
\usage{
validate_griddle(griddle)
}
\arguments{
\item{griddle}{griddle file contents}
}
\value{
\code{TRUE}, or error
}
\description{
Validate griddle file contents
}
